requirements.txt
numpy
xarray
pyserial
opencv-python
pymba
numpngw
Pillow
